requirements.txt
bta-lib==1.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dash==2.8.0
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.1
pandas==1.5.3
plotly==5.13.0
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tenacity==8.1.0
urllib3==1.26.14
Werkzeug==2.2.2
